# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Aluno Online API",
        "description": "API Rest das funcionalidades do Aluno Online",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
        "/notas": {
            "description": "Notas do Aluno Online",
            "get": {
                "description": "Lista as notas de um aluno",
                "responses": {
                    "200": {
                        "description": "Operação de Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Faltas"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/faltas": {
            "description": "Faltas do Aluno Onliine",
            "get": {
                "description": "Faltas do Aluno Online",
                "responses": {
                    "200": {
                        "description": "Operação de Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Faltas"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/boletos": {
            
        },
        "/boletos/{id}": {
            "get": {
                "description": "Exibe um boleto pelo ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação de Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Boletos"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Boleto não encontrado!"
                    }
                }
            }
        },
        "/requerimentos": {},
        "/requerimentos/{id}": {}
    },
    "components": {
        "schemas": {
            "Notas": {
                "type": "object",
                "properties": {
                    "periodo": {
                        "type": "string",
                        "example": "2021/1"
                    },
                    "disciplina": {
                        "type": "object",
                        "properties": {
                            "codigo": {
                                "type": "string",
                                "example": "ADS030"
                            },
                            "nome": {
                                "type": "string",
                                "example": "Manutenção de Software e DevOps"
                            }
                        }
                    },
                    "a1": {
                        "type": "string",
                        "example": "5.2"
                    },
                    "a2": {
                        "type": "string",
                        "example": "3.8"
                    },
                    "a3": {
                        "type": "string",
                        "example": "4.8"
                    },
                    "media": {
                        "type": "string",
                        "example": "5.0"
                    }
                }
            },
            "Faltas": {
                "type": "object",
                "properties": {
                    "periodo": {
                        "type": "string",
                        "example": "2021/1"
                    },
                    "disciplina": {
                        "type": "object",
                        "properties": {
                            "codigo": {
                                "type": "string",
                                "example": "ADS030"
                            },
                            "nome": {
                                "type": "string",
                                "example": "Manutenção de Software e DevOps"
                            }
                        }
                    },
                    "total": {
                        "type": "string",
                        "example": "0"
                    },
                    "porcentagem": {
                        "type": "string",
                        "example": "0"
                    }
                }
            },
            "Boletos": {
                "type": "object",
                "properties": {
                    "mensalidade": {
                        "type": "string",
                        "example": "01"
                    },
                    "vencimento": {
                        "type": "string",
                        "example": "01/01/2022"
                    },
                    "valor": {
                        "type": "number",
                        "example": "500.00"
                    },
                    "situacao": {
                        "type": "string",
                        "example": "Pago"
                    }
                }
            }
        }
    }
}